+++ boris-core/src/Boris/Core/Data/Instance.hs
@@ -1,0 +1,24 @@
+{-# LANGUAGE NoImplicitPrelude #-}
+module Boris.Core.Data.Instance (
+    BuildInstance (..)
+  , DiscoverInstance (..)
+  ) where
+
+import           Boris.Core.Data
+import           Boris.Core.Data.Configuration
+
+import           P
+
+data BuildInstance =
+  BuildInstance {
+      buildSpecification :: Specification
+    , buildRef :: Ref
+    , buildCommit :: Commit
+    } deriving (Eq, Show)
+
+data DiscoverInstance =
+  DiscoverInstance {
+      discoverBuild :: Build
+    , discoverRef :: Ref
+    , discoverCommit :: Commit
+    } deriving (Eq, Ord, Show)
